+++ django_httpolice/common.py
@@ -8,7 +8,7 @@ DEFAULT_SETTINGS = {
     'HTTPOLICE_ENABLE': False,
     'HTTPOLICE_BACKLOG': 20,
     'HTTPOLICE_SILENCE': [1110],
-    'HTTPOLICE_RAISE': False,
+    'HTTPOLICE_RAISE': None,
 }
 
 
@@ -19,7 +19,7 @@ def get_setting(name):
 
 class ProtocolError(Exception):
 
-    """A protocol error in a response produced by the application."""
+    """A protocol problem in a response produced by the application."""
 
     def __init__(self, exchange):
         # Importing `httpolice` can execute a lot of code,
@@ -28,6 +28,6 @@ class ProtocolError(Exception):
         buf = six.BytesIO()
         httpolice.text_report([exchange], buf)
         super(ProtocolError, self).__init__(
-            u'HTTPolice found errors in this response:\n' +
+            u'HTTPolice found problems in this response:\n' +
             buf.getvalue().decode('utf-8')
         )