+++ test/test_main.py
@@ -104,13 +104,31 @@ def test_disabled(client):
 
 
 def test_raise(client):
+    with override_settings(HTTPOLICE_RAISE='error'):
+        with pytest.raises(django_httpolice.ProtocolError):
+            client.get('/api/v1/words/?query=er')
+        assert len(django_httpolice.backlog) == 1
+
+
+def test_raise_boolean(client):
+    # Test the old format of ``HTTPOLICE_RAISE`` for backward compatibility.
     with override_settings(HTTPOLICE_RAISE=True):
         with pytest.raises(django_httpolice.ProtocolError):
             client.get('/api/v1/words/?query=er')
         assert len(django_httpolice.backlog) == 1
 
 
+def test_raise_comment(client):
+    # Disable silencing notice 1110 ("<status> response with no Date header").
+    with override_settings(HTTPOLICE_SILENCE=[]):
+        with override_settings(HTTPOLICE_RAISE='error'):
+            client.get('/api/v1/words/')        # 1110 is only a comment
+        with override_settings(HTTPOLICE_RAISE='comment'):
+            with pytest.raises(django_httpolice.ProtocolError):
+                client.get('/api/v1/words/')
+
+
 def test_no_raise_when_silenced(client):
-    with override_settings(HTTPOLICE_RAISE=True):
+    with override_settings(HTTPOLICE_RAISE='error'):
         client.get('/api/v1/words/?query=er',
                    HTTP_HTTPOLICE_SILENCE='1038 resp')